Add pizza fire
This commit adds the pizza fire mechanics. To decide in which direction
pizza should be fired, I've used ray fire technique. Basically a ray,
starting from camera's position and travelling towards current mouse
position is checked against ground and all the other collidable
entities. If this ray hits any of the collidables or ground and if
left mouse button is pressed, a new instance of pizza will be fired.
Right now all the pizzas keep on going indefinitely and needs to be
fixed.

Also, camera shake has been modified slightly. It now moves the target
instead of position.